comment

+++ src/dao/impl/TblDetailUserJapanDaoImpl.java
@@ -30,28 +30,30 @@ public class TblDetailUserJapanDaoImpl extends BaseDaoImpl implements TblDetailU
 			throws SQLException {
 		Boolean isSuccess = false;
 		try {
-			StringBuffer sql = new StringBuffer();
-			int total = 0;
-			sql.append("insert into tbl_detail_user_japan");
-			sql.append("(user_id,code_level,start_date,end_date,total) ");
-			sql.append("values(?,?,?,?,?) ");
-			SimpleDateFormat dt1 = new SimpleDateFormat("yyyy-MM-dd");
-			PreparedStatement ps = null;
-			int i = 0;
-			ps = connection.prepareStatement(sql.toString());
-			ps.setInt(++i, tblDetailUserJapan.getUserId());
-			ps.setString(++i, tblDetailUserJapan.getCodeLevel());
-			//ps.setDate(++i, (Date) tblDetailUserJapan.getStartDate());
-			ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getStartDate())));
-			//ps.setDate(++i, (Date) tblDetailUserJapan.getEndDate());
-			ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getEndDate())));
-			if (!"".equals(tblDetailUserJapan.getTotal()) || tblDetailUserJapan.getTotal() != null) {
-				ps.setInt(++i, Integer.parseInt(tblDetailUserJapan.getTotal()));
-			} else {
-				ps.setInt(++i, total);
+			if (connection != null) {
+				StringBuffer sql = new StringBuffer();
+				SimpleDateFormat dt1 = new SimpleDateFormat("yyyy-MM-dd");
+				PreparedStatement ps = null;
+				// biến set giá trị cho total của detail user japan
+				int total = 0;
+				// biến set param cho prepared statement
+				int i = 0;
+				sql.append("insert into tbl_detail_user_japan");
+				sql.append("(user_id,code_level,start_date,end_date,total) ");
+				sql.append("values(?,?,?,?,?) ");
+				ps = connection.prepareStatement(sql.toString());
+				ps.setInt(++i, tblDetailUserJapan.getUserId());
+				ps.setString(++i, tblDetailUserJapan.getCodeLevel());
+				ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getStartDate())));
+				ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getEndDate())));
+				if (!"".equals(tblDetailUserJapan.getTotal()) || tblDetailUserJapan.getTotal() != null) {
+					ps.setInt(++i, Integer.parseInt(tblDetailUserJapan.getTotal()));
+				} else {
+					ps.setInt(++i, total);
+				}
+				// update thành công thì biến isSuccess bằng true
+				isSuccess = ps.executeUpdate() > 0;
 			}
-
-			isSuccess = ps.executeUpdate() > 0;
 		} catch (SQLException e) {
 			e.printStackTrace();
 			throw e;
@@ -65,29 +67,34 @@ public class TblDetailUserJapanDaoImpl extends BaseDaoImpl implements TblDetailU
 	@Override
 	public Boolean updateDetailUserJapan(TblDetailUserJapan tblDetailUserJapan, Connection connection)
 			throws SQLException {
-		StringBuffer sql = new StringBuffer();
-		Boolean isSuccess = false;
-		int total = 0;
-		sql.append("update tbl_detail_user_japan ");
-		sql.append("set code_level = ?,start_date = ?,end_date = ?,total = ? ");
-		sql.append("where user_id = ?");
-		SimpleDateFormat dt1 = new SimpleDateFormat("yyyy-MM-dd");
-		PreparedStatement ps = null;
+		Boolean isSuccess = false;	
 		try {
-			ps = connection.prepareStatement(sql.toString());
-			int i = 0;
-			ps.setString(++i, tblDetailUserJapan.getCodeLevel());
-			ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getStartDate())));
-			ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getEndDate())));
-			if (tblDetailUserJapan.getTotal() != "" || tblDetailUserJapan.getTotal() != null) {
-				ps.setInt(++i, Integer.parseInt(tblDetailUserJapan.getTotal()));
-			} else {
-				ps.setInt(++i, total);
+			if (connection != null) {
+				StringBuffer sql = new StringBuffer();
+				PreparedStatement ps = null;
+				SimpleDateFormat dt1 = new SimpleDateFormat("yyyy-MM-dd");
+				// biến set giá trị cho total của detail user japan
+				int total = 0;
+				// biến set param cho prepared statement
+				int i = 0;
+				
+				sql.append("update tbl_detail_user_japan ");
+				sql.append("set code_level = ?,start_date = ?,end_date = ?,total = ? ");
+				sql.append("where user_id = ?");
+				ps = connection.prepareStatement(sql.toString());
+				ps.setString(++i, tblDetailUserJapan.getCodeLevel());
+				ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getStartDate())));
+				ps.setDate(++i, java.sql.Date.valueOf(dt1.format(tblDetailUserJapan.getEndDate())));
+				if (tblDetailUserJapan.getTotal() != "" || tblDetailUserJapan.getTotal() != null) {
+					ps.setInt(++i, Integer.parseInt(tblDetailUserJapan.getTotal()));
+				} else {
+					ps.setInt(++i, total);
+				}
+				if (tblDetailUserJapan.getUserId() != 0) {
+					ps.setInt(++i, tblDetailUserJapan.getUserId());
+				}
+				isSuccess = ps.executeUpdate() > 0;
 			}
-			if (tblDetailUserJapan.getUserId() != 0) {
-				ps.setInt(++i, tblDetailUserJapan.getUserId());
-			}
-			isSuccess = ps.executeUpdate() > 0;
 		} catch (SQLException e) {
 			e.printStackTrace();
 			throw e;
@@ -102,13 +109,13 @@ public class TblDetailUserJapanDaoImpl extends BaseDaoImpl implements TblDetailU
 	public TblDetailUserJapan getDetailUserJapanById(int userId) {
 		TblDetailUserJapan tblUserJapan = new TblDetailUserJapan();
 		try {
-			StringBuffer sql = new StringBuffer();
-			sql.append("select code_level from tbl_detail_user_japan ");
-			sql.append("where user_id = ?");
-			PreparedStatement ps = null;
-			ResultSet rs = null;
 			connectDB();
 			if (conn != null) {
+				StringBuffer sql = new StringBuffer();
+				PreparedStatement ps = null;
+				ResultSet rs = null;
+				sql.append("select code_level from tbl_detail_user_japan ");
+				sql.append("where user_id = ?");
 				ps = conn.prepareStatement(sql.toString());
 				int i = 0;
 				ps.setInt(++i, userId);
@@ -124,24 +131,26 @@ public class TblDetailUserJapanDaoImpl extends BaseDaoImpl implements TblDetailU
 		}
 		return tblUserJapan;
 	}
-	
+
 	/**
 	 * @see dao.TblDetailUserJapanDao#deleteDetailUserJapan(int,Connection)
 	 */
 	@Override
 	public Boolean deleteDetailUserJapan(int userId, Connection connection) throws SQLException {
-		StringBuffer sql = new StringBuffer();
 		Boolean isSuccess = false;
-		sql.append("delete from tbl_detail_user_japan ");
-		sql.append("where user_id = ?");
-		PreparedStatement ps = null;
 		try {
-			ps = connection.prepareStatement(sql.toString());
-			int i = 0;
-			if (userId != 0) {
-				ps.setInt(++i, userId);
+			if (connection != null) {
+				StringBuffer sql = new StringBuffer();
+				PreparedStatement ps = null;
+				sql.append("delete from tbl_detail_user_japan ");
+				sql.append("where user_id = ?");
+				ps = connection.prepareStatement(sql.toString());
+				int i = 0;
+				if (userId != 0) {
+					ps.setInt(++i, userId);
+				}
+				isSuccess = ps.executeUpdate() > 0;
 			}
-			isSuccess = ps.executeUpdate() > 0;
 		} catch (SQLException e) {
 			throw e;
 		}
